requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
openai
faiss-cpu
shap
fairlearn
pytest
langchain-community
langchain-openai
